YellowTech Admin Page
First I have worked on the login and register forms for a user to be able to properly log in sign out and register. I have changed the design and used what was required. Furthermore, I have made some changes in the layouts folder for: _sidebar and _header where unnecessary code was removed and  a Categories link an Order one a Users one additionally to the Dashboard link were added. The Users page was created to show users in a table as requested with the delete and edit actions. Also I have filled the code for the new.blade.php to add a new user, the edit.blade.php to edit a user . All was made with the design of AdminLTE from docs. And the necessary controllers like AuthController, AdminController, and so on  were edited to accomplish login registration edit delete and update actions.